SplitLearnComm/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "splitlearn-comm"
version = "1.0.0"
description = "A gRPC-based communication library for distributed computing"
authors = [
    {name = "SplitLearn Contributors"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = ["grpc", "distributed-computing", "deep-learning", "communication", "rpc"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing",
]

dependencies = [
    "torch>=2.0.0",
    "grpcio>=1.50.0",
    "protobuf>=4.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "grpcio-tools>=1.50.0",  # For compiling protobuf
]

ui = [
    "gradio>=3.50.0,<5.0.0",  # Gradio UI framework
    "pandas>=1.5.0",  # For server monitoring data handling
]

all = [
    "splitlearn-comm[dev,ui]",
]

[project.urls]
Homepage = "https://github.com/yourusername/SplitLearnComm"
Repository = "https://github.com/yourusername/SplitLearnComm"
Issues = "https://github.com/yourusername/SplitLearnComm/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
